categories back

+++ core_service/app/handlers/event.py
@@ -78,23 +78,27 @@ class EventHandler(EventService, metaclass=WorkerMetaBase):
     """Class contains method for handling event stuff."""
 
     def create_event(
-        self, user_id, account_id, event_time,
-        diff, description
+        self,
+        user_id,
+        account_id,
+        category_id,
+        event_time,
+        diff,
+        description
     ):
         """Create new event."""
-        accounts_response = AccountHandler.get_accounts(user_id)
+        accounts_response = AccountHandler.check_account_user(
+            account_id,
+            user_id
+        )
         if accounts_response['status'] != 'OK':
             return accounts_response
-        if not any(
-            user_acc['id'] == account_id
-            for user_acc in accounts_response['accounts']
-        ):
-            return {'status': 'no such account'}
 
         event_time = datetime.fromtimestamp(event_time)
         event = EventModel(
             user_id=user_id,
             account_id=account_id,
+            category_id=category_id,
             event_time=event_time,
             diff=diff,
             description=description
@@ -121,6 +125,13 @@ class EventHandler(EventService, metaclass=WorkerMetaBase):
 
         If 0 filters provided, get some event on account.
         """
+        accounts_response = AccountHandler.check_account_user(
+            account_id,
+            user_id
+        )
+        if accounts_response['status'] != 'OK':
+            return accounts_response
+
         if start_time is not None:
             start_time = datetime.fromtimestamp(start_time)
         if end_time is not None:
@@ -138,15 +149,25 @@ class EventHandler(EventService, metaclass=WorkerMetaBase):
         }
 
     def get_events(
-        self, user_id, account_id,
-        start_time=None, end_time=None,
-        label=None
+        self,
+        user_id,
+        account_id,
+        start_time=None,
+        end_time=None,
+        category_id=None
     ):
         """
         Get all events user has.
 
         If 0 filters provided, get every event on account.
         """
+        accounts_response = AccountHandler.check_account_user(
+            account_id,
+            user_id
+        )
+        if accounts_response['status'] != 'OK':
+            return accounts_response
+
         if start_time is not None:
             start_time = datetime.fromtimestamp(start_time)
         if end_time is not None:
@@ -158,23 +179,44 @@ class EventHandler(EventService, metaclass=WorkerMetaBase):
             query = query.filter(EventModel.event_time > start_time)
         if end_time:
             query = query.filter(EventModel.event_time < end_time)
+        if category_id is not None:
+            query = query.filter(EventModel.category_id == category_id)
         return {
             'status': 'OK',
             'events': [event_schema.dump(event) for event in query.all()]
         }
 
-    def edit_event(self, user_id, event_id, event_time, diff, description):
+    def edit_event(
+        self,
+        user_id,
+        account_id,
+        event_id,
+        category_id,
+        event_time,
+        diff,
+        description
+    ):
         """Edit existing event."""
+        accounts_response = AccountHandler.check_account_user(
+            account_id,
+            user_id
+        )
+        if accounts_response['status'] != 'OK':
+            return accounts_response
+
         event = session.get(EventModel, event_id)
         if event is None:
             return {'status': 'no such event'}
         if event.user_id != user_id:
             return {'status': 'accessing another users events'}
+        if event.account_id != account_id:
+            return {'status': 'wrong account for event'}
 
         event_time = datetime.fromtimestamp(event_time)
         old_event_time: datetime = event.event_time
         old_diff = event.diff
 
+        event.category_id = category_id
         event.event_time = event_time
         event.diff = diff
         event.description = description
@@ -207,13 +249,22 @@ class EventHandler(EventService, metaclass=WorkerMetaBase):
         session.commit()
         return {'status': 'OK', 'event': event_schema.dump(event)}
 
-    def delete_event(self, user_id, event_id):
+    def delete_event(self, user_id, account_id, event_id):
         """Delete existing event."""
+        accounts_response = AccountHandler.check_account_user(
+            account_id,
+            user_id
+        )
+        if accounts_response['status'] != 'OK':
+            return accounts_response
+
         event = session.get(EventModel, event_id)
         if event is None:
             return {'status': 'no such event'}
         if event.user_id != user_id:
             return {'status': 'accessing another users events'}
+        if event.account_id != account_id:
+            return {'status': 'wrong account for event'}
 
         session.delete(event)
         # TODO remove savepoint if event is a last one
@@ -228,6 +279,13 @@ class EventHandler(EventService, metaclass=WorkerMetaBase):
 
     def get_balance(self, user_id, account_id, timestamp):
         """Get balance on given account in given point in time."""
+        accounts_response = AccountHandler.check_account_user(
+            account_id,
+            user_id
+        )
+        if accounts_response['status'] != 'OK':
+            return accounts_response
+
         timepoint = datetime.fromtimestamp(timestamp)
         savepoint = session\
             .query(SavePointModel)\
